ядро обновлено до SciTE 5.6.2 исправлен lpeg паттерн для ini и props файлов для построения списка функций в боковой панели добавлена переведённая документация для Lua 5.5 исправлена система событий в случае использования корутин скрипт Text Highlighting подсвечивает все выделенные блоки, если их несколько убрана возможность установки закладки для несуществующго на накопителе файла

+++ scite/src/ExportPDF.cxx
@@ -326,7 +326,7 @@ void SciTEBase::SaveToPDF(const FilePath &saveName) {
 			const double glyphWidth = fontToPoints(PDFfontWidths[fontSet]);
 			xPos += glyphWidth;
 			// if cannot fit into a line, flush, wrap to next line
-			if (xPos > pageWidth - pageMargin.right) {
+			if (xPos > static_cast<double>(pageWidth - pageMargin.right)) {
 				nextLine();
 				xPos += glyphWidth;
 			}
@@ -365,7 +365,7 @@ void SciTEBase::SaveToPDF(const FilePath &saveName) {
 			firstLine = true;
 			pageCount++;
 			const double fontAscender = fontToPoints(PDFfontAscenders[fontSet]);
-			yPos = pageHeight - pageMargin.top - fontAscender;
+			yPos = static_cast<double>(pageHeight - pageMargin.top) - fontAscender;
 			// start a new page
 			snprintf(buffer, std::size(buffer), "BT 1 0 0 1 %d %d Tm\n",
 				pageMargin.left, (int)yPos);
